// src/entities/representante/model/representante.types.ts
export interface IRepresentanteRow {
  id: string;
  nombres: string | null;
  apellidos: string | null;
  nombre: string;
  cedula: string | null;
  celular: string | null;
  email: string | null;
  direccion: string | null;
  ocupacion: string | null;
  perfil_id: string | null;
  hijos: number;
}